requirements.txt
python-telegram-bot==20.6
aiohttp
requests
beautifulsoup4
nest_asyncio
